// sourcecode/BTrak.Packages/snova-admin-module/admin-module-components/admin-components/src/lib/admin-module/models/projects/testcasedropdown.ts
export class TestCaseDropdownList {
    id: string;
    userId: string;
    value: string;
    fullName: string;
    description: string;
    isArchived: boolean;
    sectionId: string;
    auditCategoryId: string;
    statusHexValue: string;
    statusShortName: string;
    profileImage: string;
    isHierarchical: boolean;
    timeStamp: any;
}